refactor(app): migrate copy/delete/head object usecases
